+++ packages/basic-layouts/src/Menu/index.tsx
@@ -1,14 +1,42 @@
 import React, { Fragment } from 'react';
 import classnames from 'classnames';
 import Menu, { MenuItemProps, SubMenuProps } from '@uiw/react-menu';
-import { DefaultProps } from '@uiw-admin/router-control';
+import {
+  DefaultProps,
+  RoutesBaseProps,
+  Routers,
+} from '@uiw-admin/router-control';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { matchPath, NavigateFunction } from 'react-router';
+import { getRoutesList } from './../utils';
+import { SearchSelect, SearchSelectOptionData } from 'uiw';
+
+import pinyin from 'pinyin';
+
+type ValueType = string | number;
 
 interface MenuProps {
   collapsed?: boolean;
   routes?: DefaultProps['routes'];
 }
+
+const onNavigate = (
+  item: Omit<Routers, 'navigate'> & { navigate: Function | string },
+  navigate: NavigateFunction,
+) => {
+  if (Reflect.has(item, 'navigate') && item.navigate) {
+    if (typeof item.navigate === 'function') {
+      item.navigate(navigate);
+      return;
+    }
+    const Fun = new Function(`return ${item.navigate}`)();
+    if (typeof Fun === 'function') {
+      Fun(navigate);
+    }
+    return;
+  }
+  return true;
+};
 
 function renderMenuItem(
   routes: MenuProps['routes'] = [],
@@ -37,6 +65,7 @@ function renderMenuItem(
     if (matchPath({ path: item.path }, pathName)) {
       props.active = true;
     }
+
     if (item.routes) {
       if (collapsed) {
         props.overlayProps = {
@@ -44,6 +73,14 @@ function renderMenuItem(
           usePortal: true,
         };
       }
+      if (
+        item.routes.find((route: RoutesBaseProps) => route.path === pathName)
+      ) {
+        props.overlayProps = {
+          isOpen: true,
+        };
+      }
+
       return (
         <Menu.SubMenu {...props} text={item.name || '-'} collapse={collapsed}>
           {renderMenuItem(item.routes, collapsed, pathName, navigate)}
@@ -54,15 +91,8 @@ function renderMenuItem(
       <Menu.Item
         {...props}
         onClick={() => {
-          if (Reflect.has(item, 'navigate') && item.navigate) {
-            if (typeof item.navigate === 'function') {
-              item.navigate(navigate);
-              return;
-            }
-            const Fun = new Function(`return ${item.navigate}`)();
-            if (typeof Fun === 'function') {
-              Fun(navigate);
-            }
+          const result = onNavigate(item, navigate);
+          if (!result) {
             return;
           }
           navigate(item.path, { replace: true });
@@ -74,18 +104,127 @@ function renderMenuItem(
   });
 }
 
+const SearchMenus = (props: MenuProps) => {
+  const { routes = [] } = props || {};
+
+  const navigate = useNavigate();
+
+  const listRouters = React.useMemo(() => {
+    return getRoutesList(routes);
+  }, [routes]);
+
+  const listMenus = React.useMemo(() => {
+    return listRouters
+      .filter((item) => {
+        return (
+          item &&
+          'name' in item &&
+          !item.hideInMenu &&
+          !item.index &&
+          item.path !== '*'
+        );
+      })
+      .map((item) => {
+        let pyInitials = pinyin(item.name || '', {
+          style: pinyin.STYLE_FIRST_LETTER,
+        });
+
+        let py = pinyin(item.name || '', {
+          style: pinyin.STYLE_NORMAL,
+        });
+
+        let pyInitialsValue = '';
+        let pyValue = '';
+
+        if (Array.isArray(py)) {
+          pyValue = py.join('');
+        }
+
+        if (Array.isArray(pyInitials)) {
+          pyInitialsValue = pyInitials.join('');
+        }
+        return {
+          label: item.name,
+          value: item.path,
+          py: pyValue,
+          pyInitials: pyInitialsValue,
+        } as SearchSelectOptionData;
+      });
+  }, [listRouters]);
+
+  const [list, setList] = React.useState<SearchSelectOptionData[]>(listMenus);
+
+  const onChange = (
+    event: SearchSelectOptionData[] | ValueType | ValueType[],
+  ) => {
+    if (Array.isArray(event)) {
+      const [obj] = event;
+      if (obj && typeof obj === 'object' && obj.value) {
+        const current: any = listRouters.find(
+          (item) => item.path === obj.value,
+        );
+        console.log(current);
+        const result = onNavigate(current, navigate);
+        if (!result) {
+          return;
+        }
+        navigate(current.path, { replace: true });
+      }
+    }
+  };
+
+  const handleSearch = (value: string) => {
+    if (value) {
+      const li = listMenus.filter((item) => {
+        const Reg = new RegExp(`${value}`);
+        if (
+          Reg.test(item.label) ||
+          Reg.test(item.py) ||
+          Reg.test(item.pyInitials)
+        ) {
+          return true;
+        }
+        return false;
+      });
+      setList(li);
+    } else {
+      setList(listMenus);
+    }
+  };
+
+  return (
+    <div style={{ marginBottom: 10, padding: '0px 10px' }}>
+      <SearchSelect
+        placeholder="请输入"
+        mode="single"
+        labelInValue
+        showSearch={true}
+        option={list}
+        onSearch={handleSearch}
+        onChange={onChange}
+      />
+    </div>
+  );
+};
 export default (props: MenuProps = {}) => {
   const { routes = [], collapsed } = props;
   const location = useLocation();
   const navigate = useNavigate();
   const pathName = location.pathname;
+
+  // @ts-ignore
+  const searchMenu = SEARCH_MENU && <SearchMenus routes={routes} />;
+
   return (
-    <Menu
-      theme="dark"
-      inlineCollapsed={!!collapsed}
-      style={{ padding: '0 12px' }}
-    >
-      {renderMenuItem(routes, !!collapsed, pathName, navigate, true)}
-    </Menu>
+    <React.Fragment>
+      {searchMenu}
+      <Menu
+        theme="dark"
+        inlineCollapsed={!!collapsed}
+        style={{ padding: '0 12px' }}
+      >
+        {renderMenuItem(routes, !!collapsed, pathName, navigate, true)}
+      </Menu>
+    </React.Fragment>
   );
 };