+++ packages/components/src/ProTable/BaseForm.tsx
@@ -13,13 +13,13 @@ import {
   MonthPicker,
   FormSubmitProps,
   SearchSelect,
-  // TreeChecked
+  FormRefType,
 } from 'uiw';
 import Select from './widgets/Select';
 import FormRadio from './widgets/Radio';
 import SearchTree from './widgets/SearchTree';
 import { useStore } from './hooks';
-import { BaseFormProps, Fields, FormProps } from './types';
+import { BaseFormProps, Fields, FormProps, FormCol } from './types';
 
 const widgets = {
   input: Input,
@@ -39,11 +39,11 @@ const widgets = {
 const BaseForm: React.FC<BaseFormProps> = (props) => {
   const store = useStore();
 
-  const formRef = useRef<any>();
+  const formRef = useRef<FormRefType>();
 
   let { updateStore, updateForm } = store as any;
 
-  const { columns, searchBtns, onBeforeSearch } = props;
+  const { columns, searchBtns, onBeforeSearch, formCol = 5 } = props;
 
   const getFields = (formProps: FormProps, title: any) => {
     const { widgetProps, key, widget, label, initialValue, ...otherProps } =
@@ -60,21 +60,28 @@ const BaseForm: React.FC<BaseFormProps> = (props) => {
   // 获取表单配置
   const getFormFields = useMemo(() => {
     const fields: Fields = {};
-    columns.forEach((col) => {
-      if (col.props && Object.keys(col.props).length > 0) {
-        if (Array.isArray(col.props)) {
-          col.props.forEach((f) => {
-            const name = f.key || col.key;
-            fields[name] = getFields(f, col.title);
-          });
-        } else {
-          const { key } = col.props;
-          const name = key || col.key;
-          fields[name] = getFields(col.props, col.title);
+    const helper = (cols: FormCol[]) => {
+      cols.forEach((col) => {
+        if (col.props && Object.keys(col.props).length > 0) {
+          if (Array.isArray(col.props)) {
+            col.props.forEach((f) => {
+              const name = f.key || col.key;
+              fields[name] = getFields(f, col.title);
+            });
+          } else {
+            const { key } = col.props;
+            const name = key || col.key;
+            fields[name] = getFields(col.props, col.title);
+          }
         }
-      }
-    });
+        // 多层children
+        if (col.children && col.children.length > 0) {
+          helper(col.children);
+        }
+      });
+    };
 
+    helper(columns);
     return fields;
   }, [JSON.stringify(columns)]);
 
@@ -95,8 +102,11 @@ const BaseForm: React.FC<BaseFormProps> = (props) => {
     }
   }, [formRef]);
 
+  // 表单占比
+  const cent = 100 / formCol + '%';
+
   const itemsLength = Object.keys(getFormFields).length;
-  const emptyLength = 4 - (itemsLength % 5);
+  const emptyLength = formCol - 1 - (itemsLength % formCol);
 
   return (
     <Form
@@ -126,15 +136,13 @@ const BaseForm: React.FC<BaseFormProps> = (props) => {
           <div>
             <Row gutter={12}>
               {Object.keys(fields).map((key) => (
-                <Col key={key} fixed style={{ width: '20%' }}>
+                <Col key={key} fixed style={{ width: cent }}>
                   {fields[key]}
                 </Col>
               ))}
-              {Array(emptyLength)
-                .fill('')
-                .map((value, index) => (
-                  <Col key={index.toString()} fixed style={{ width: '20%' }} />
-                ))}
+              {Array.from({ length: emptyLength }, (_, index) => (
+                <Col key={index.toString()} fixed style={{ width: cent }} />
+              ))}
               <Col
                 align="bottom"
                 style={{