+++ packages/components/src/ProTable/BaseTable.tsx
@@ -46,19 +46,11 @@ const BaseTable: React.FC<BaseTableProps> = ({
   const { x } = scroll;
   const isCheckbox = type === 'checkbox';
 
-  // columns列和标题是否居中
+  // 提出表单属性
   const defaultColumns = () => {
     return columns.map((item) => {
-      const { align = 'left', render } = item;
-      return {
-        ...item,
-        // 标题样式
-        style: { textAlign: align },
-        // 列内容
-        render: render
-          ? render
-          : (text: string) => <div style={{ textAlign: align }}>{text}</div>,
-      };
+      const { props, ...others } = item;
+      return others;
     });
   };
 
@@ -137,7 +129,11 @@ const BaseTable: React.FC<BaseTableProps> = ({
     defaultSelected,
     type === 'radio',
   );
-
+  //  总数
+  const total =
+    formatData && data
+      ? formatData(data).total
+      : data?.total || prevData?.total;
   // 分页
   const onPageChange = useCallback(
     async (page) => {
@@ -168,8 +164,8 @@ const BaseTable: React.FC<BaseTableProps> = ({
       }
     });
     const stores: any = {
-      data: data?.data,
-      total: data?.total,
+      data: tableData,
+      total,
       loading: isValidating,
       // onSearch,
       selection,
@@ -187,16 +183,17 @@ const BaseTable: React.FC<BaseTableProps> = ({
 
     // 上一次请求数据
     if (data) {
-      setPrevData(formatData ? formatData(data) : data);
+      setPrevData(tableData);
     }
   }, [
-    JSON.stringify(data),
+    JSON.stringify(tableData),
     isValidating,
     JSON.stringify(columns),
     pageIndex,
     JSON.stringify(selection),
     setPageIndex,
     mutate,
+    total,
   ]);
 
   const selectionCol = [
@@ -257,11 +254,7 @@ const BaseTable: React.FC<BaseTableProps> = ({
                 {...paginationProps}
                 current={pageIndex}
                 pageSize={pageSize}
-                total={
-                  formatData && data
-                    ? formatData(data).total
-                    : data?.total || prevData?.total
-                }
+                total={total}
                 onChange={(page) => {
                   onPageChange(page);
                 }}